little bug in gradle-find-project-dir, it wasn't returning a path with a slash at the end.
This was TECHNICALLY finding the right directory, but when run from
compile (which displayed the right directory, sans ending "/"), for some
reason, gradle was trying to run from one more directory up.

Obviously, this isn't cool.  But it should be fixed now.

Tests reflect that it needs this slash as well.