better defined integration tests -- found `compilation-directory` and `compile-command` variables that I am able to check instead of trying to capture my own.
No longer need to mock, just had to set yes-or-no-p to always be true so
the tests wouldn't interrupt.  Makes the tests cleaner.

The gradle executable path is just a stub anyway, so the compilation
won't happen because that command is off.  This is just set this way so
I can check the compilation-command is what is expected.

The compilation-directory is in accordance with the unit tests as well,
checking for the slash at the end of the path.