support for gradlew, optional for the user.
Changed to have a function run-from-dir, which retruns the directory to
run commands from.  This is either a directory with a gradle build file
directory (eg. build.gradle) or a directory with a gradlew file.

Two predicate functions broken out to establish criteria for one of
these directories.

Introduced a boolean `gradle-use-gradlew`, and do the if checks to pick
the right predicate function and pick the right executable name (gradlew
or gradle).

New feature to break out the scenarios so they aren't cluttered.

Fixed a few lint errors as well.